Moved stuff, again
 Changes to be committed:
	modified:   src/main/java/mounderfod/moundertweaks/MounderTweaksMain.java
	modified:   src/main/java/mounderfod/moundertweaks/mixin/client/OptionScreenMixin.java
	modified:   src/main/java/mounderfod/moundertweaks/util/MounderTweaksModmenuImpl.java
	renamed:    src/main/java/mounderfod/moundertweaks/util/ClientConfig.java -> src/main/java/mounderfod/moundertweaks/util/config/ClientConfig.java
	renamed:    src/main/java/mounderfod/moundertweaks/util/CommonConfig.java -> src/main/java/mounderfod/moundertweaks/util/config/CommonConfig.java
	renamed:    src/main/java/mounderfod/moundertweaks/util/MounderTweaksConfig.java -> src/main/java/mounderfod/moundertweaks/util/config/MounderTweaksConfig.java
