+++ src/main/java/mounderfod/moundertweaks/mixin/client/OptionScreenMixin.java
@@ -1,7 +1,7 @@
 package mounderfod.moundertweaks.mixin.client;
 
 import me.sargunvohra.mcmods.autoconfig1u.AutoConfig;
-import mounderfod.moundertweaks.util.MounderTweaksConfig;
+import mounderfod.moundertweaks.util.config.MounderTweaksConfig;
 
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
