+++ src/main/java/mounderfod/moundertweaks/util/MounderTweaksModmenuImpl.java
@@ -3,6 +3,7 @@ package mounderfod.moundertweaks.util;
 import io.github.prospector.modmenu.api.ConfigScreenFactory;
 import io.github.prospector.modmenu.api.ModMenuApi;
 import me.sargunvohra.mcmods.autoconfig1u.AutoConfig;
+import mounderfod.moundertweaks.util.config.MounderTweaksConfig;
 
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
